requirements.txt
click==5.1
compass-stylesheets==0.12.6
defusedxml==0.4.1
Django==1.10.3
django-mustache==1.1.0
django-reversion==2.0.7
djangorestframework==3.5.3
https://github.com/sheppard/django-media-thumbnailer/archive/master.zip
FormEncode==2.0.0a1
html-json-forms==1.0.0
linecache2==1.0.0
Markdown==2.6.7
natural-keys==1.1.0
oauthlib==2.0.1
Pillow==3.4.2
pkg-resources==0.0.0
psycopg2==2.6.2
PyJWT==1.4.2
pyScss==1.3.5
pystache==0.5.4
python-social-auth==0.2.21
python3-openid==3.0.10
pyxform==0.9.24
PyYAML==3.12
requests==2.12.1
requests-oauthlib==0.7.0
six==1.10.0
swapper==1.0.0
traceback2==1.4.0
unicodecsv==0.14.1
unittest2==1.1.0
vera==1.0.0b2
wq==1.0.0b2
wq.app==1.0.0b2
wq.core==1.0.0b1
wq.db==1.0.0b3
wq.io==1.0.0b1
wq.start==1.0.0b3
xlrd==1.0.0
xlsconv==0.3.0
